Fix some 64bit warnings that were missed in revision 3181. Reported by Jari Hautio - jaari.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3182 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/immovable.cc
@@ -893,7 +893,9 @@ PlayerImmovable::PlayerImmovable(const Map_Object_Descr & mo_descr) :
 PlayerImmovable::~PlayerImmovable()
 {
 	if (m_workers.size())
-		log("Building::~Building: %u workers left!\n", m_workers.size());
+		log
+			("Building::~Building: %lu workers left!\n",
+			 static_cast<long unsigned int>(m_workers.size()));
 }
 
 /**
